requirements.txt
openai=1.18.0
streamlit=1.45.1
python-dotenv=1.1.0